added animation to user job actions

+++ client/src/components/UserJobs.jsx
@@ -5,6 +5,7 @@ import {Card, Container, Badge, Button} from 'react-bootstrap'
 import {FaEdit, FaTimes} from 'react-icons/fa'
 import '../css/user-jobs.css'
 import EditJobModal from './EditJobModal'
+import {motion, AnimatePresence} from 'framer-motion'
 
 const UserJobs = () => {
 
@@ -27,11 +28,13 @@ const UserJobs = () => {
 
   const deleteUserJob = async (id) => {
     try {
-      const deleteUserJob = await fetch(`/api/jobs/${id}`, {
-        method: "DELETE"
-      })
-      // userJobs.filter(job => job.job_id !== id)
-      dispatch(getUserJobs(userID))
+      if(window.confirm('Are you sure you want to delete this job?')) {
+        const deleteUserJob = await fetch(`/api/jobs/${id}`, {
+          method: "DELETE"
+        })
+        // userJobs.filter(job => job.job_id !== id)
+        dispatch(getUserJobs(userID))
+      }
     } catch (error) {
       console.error(error.message)
     }
@@ -40,26 +43,35 @@ const UserJobs = () => {
   return (
     <Container>
       <h2>Your Posted Jobs</h2>
-      {userJobs === null ? (<h2>You have no jobs posted</h2>) : userJobs.map((job) => (
-        <Card key={job.job_id} className="mt-5 mb-5 p-5">
-          <h1>{job.job_title}</h1>
-          <h3>{job.company_name}</h3>
-          <h6>
-            {job.work_address} {job.city} {job._state} {job.zip}
-          </h6>
-          <div className="user-job-salary-container"><h6>Salary: </h6><Badge>{job.salary}</Badge></div>
-          <div className="user-job-benefits-container"><h6>Benefits: </h6><Badge>None right now</Badge></div>
-          <div className="user-job-type-container"><h6>Job Type: </h6><Badge>{job.job_type}</Badge></div>
-          <div className="user-job-location-container"><h6>Job Location: </h6><Badge>{job.job_location}</Badge></div>
+        <AnimatePresence>
+          {userJobs === null ? (<h2>You have no jobs posted</h2>) : userJobs.map((job) => (
+            <motion.div 
+              key={job.job_id}
+              initial={{opacity: 0}}
+              animate={{opacity: 1}}
+              exit={{opacity: 0}}
+            >
+              <Card key={job.job_id} className="mt-5 mb-5 p-5">
+                <h1>{job.job_title}</h1>
+                <h3>{job.company_name}</h3>
+                <h6>
+                  {job.work_address} {job.city} {job._state} {job.zip}
+                </h6>
+                <div className="user-job-salary-container"><h6>Salary: </h6><Badge>{job.salary}</Badge></div>
+                <div className="user-job-benefits-container"><h6>Benefits: </h6><Badge>None right now</Badge></div>
+                <div className="user-job-type-container"><h6>Job Type: </h6><Badge>{job.job_type}</Badge></div>
+                <div className="user-job-location-container"><h6>Job Location: </h6><Badge>{job.job_location}</Badge></div>
 
-          <h6>Job Description: </h6>
-          <p>{job._description}</p>
-          <div className="user-job-actions-container">
-            <Button variant="warning" onClick={handleShow}><FaEdit/> Edit</Button>
-            <Button variant="danger" onClick={() => deleteUserJob(job.job_id)}><FaTimes/> Delete</Button>
-          </div>
-        </Card>
-      ))}
+                <h6>Job Description: </h6>
+                <p>{job._description}</p>
+                <div className="user-job-actions-container">
+                  <Button variant="warning" onClick={handleShow}><FaEdit/> Edit</Button>
+                  <Button variant="danger" onClick={() => deleteUserJob(job.job_id)}><FaTimes/> Delete</Button>
+                </div>
+              </Card>
+            </motion.div>
+          ))}
+        </AnimatePresence>
       <EditJobModal show={show} handleClose={handleClose}/>
     </Container>
   )
